seprated the chart to its own component
